requirements.txt

python-telegram-bot==20.7
requests==2.31.0
phonenumbers==8.13.29
Pillow==10.2.0
python-dotenv==1.0.1
